fix(knowledge): 统一默认入口并保留独立仪表盘;
