+++ apps/negentropy-ui/app/knowledge/pipelines/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function PipelinesRedirect() {
-  redirect("/knowledge");
+  redirect("/knowledge/dashboard");
 }
